requirements.txt
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.3.15
PyYAML==5.3.1
Flask==1.1.1
boto3==1.9.82
pandas==0.23.4
PyMySQL==0.9.3
scikit-learn==0.21.2
numpy==1.17.2
seaborn==0.9.0
matplotlib==3.1.3
pickleshare==0.7.5
pytest==5.4.1